Route WebRender pref changes through the RenderBackend.
The overlap between debug flags and debug commands is a bit of a mess,
and this patch does not entirely clean it up. But this at least means
that the RenderBackend can react to pref changes in Firefox.

Differential Revision: https://phabricator.services.mozilla.com/D11958
